Code Review Chagnes, getgauge/gauge-python#33
